requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.55.3
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pillow==11.1.0
playsound==1.3.0
psutil==6.1.1
pygame==2.6.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-vlc==3.0.21203
pyttsx3==2.98
requests==2.32.3
six==1.17.0
urllib3==2.3.0
